Fix: PR 리뷰 사항 수정
- 포멧 함수 response dto로 이동
- earthDistanceKm 서비스 단에서 셋팅
