+++ src/main/java/com/dnd/runus/application/scale/ScaleService.java
@@ -6,6 +6,8 @@ import com.dnd.runus.presentation.v1.scale.dto.ScaleSummaryResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import static com.dnd.runus.global.constant.ScaleConstant.DISTANCE_KM_AROUND_THE_EARTH;
+
 @Service
 @RequiredArgsConstructor
 public class ScaleService {
@@ -16,6 +18,6 @@ public class ScaleService {
         ScaleSummary summary = scaleRepository.getSummary();
 
         return new ScaleSummaryResponse(
-                summary.totalCourseCnt(), summary.totalCourseDistanceKm(), summary.earthDistanceKm());
+                summary.totalCourseCnt(), summary.totalCourseDistanceKm(), DISTANCE_KM_AROUND_THE_EARTH);
     }
 }
