+++ src/main/java/com/dnd/runus/presentation/v1/scale/dto/ScaleSummaryResponse.java
@@ -2,6 +2,8 @@ package com.dnd.runus.presentation.v1.scale.dto;
 
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.text.NumberFormat;
+import java.util.Locale;
 
 public record ScaleSummaryResponse(
     @Schema(description = "전체 코스 수", example = "18코스")
@@ -11,4 +13,19 @@ public record ScaleSummaryResponse(
     @Schema(description = "지구 한 바퀴 거리", example = "40,075km")
     String earthDistanceKm
 ) {
+    public ScaleSummaryResponse(int totalCourseCnt, int totalCourseDistanceMeter, String earthDistanceKm) {
+        this(courseCountFormater(totalCourseCnt), meterToKM(totalCourseDistanceMeter), earthDistanceKm);
+    }
+
+    private static String meterToKM(int totalCourseDistanceMeter) {
+        double kilometers = totalCourseDistanceMeter / 1000.0;
+
+        NumberFormat numberFormat = NumberFormat.getInstance(Locale.US);
+        numberFormat.setGroupingUsed(true); // 쉼표 구분 사용
+        return numberFormat.format(kilometers) + "km";
+    }
+
+    private static String courseCountFormater(int totalCourseCnt) {
+        return totalCourseCnt + "코스";
+    }
 }
